requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
blessings==1.7
bpython==0.22.1
certifi==2021.10.8
charset-normalizer==2.0.7
conda==4.3.16
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.10
cwcwidth==0.1.5
distlib==0.3.3
Django==2.2.10
django-cors-headers==3.10.1
django-extensions==3.1.5
django-nested-inline==0.4.4
django-utils-six==2.0
djangorestframework==3.9.4
drf-yasg==1.17.1
filelock==3.4.0
future==0.18.2
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
pipenv==2021.11.15
platformdirs==2.4.0
psycopg2-binary==2.8.6
pycosat==0.6.3
Pygments==2.10.0
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
pyxdg==0.27
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
